requirements.txt
alembic==1.13.3
fastapi==0.115.0
httpx==0.27.2
psycopg2==2.9.9
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==1.4.46
uvicorn==0.30.6
